+++ src/i18n_check/check/invalid_keys.py
@@ -26,7 +26,7 @@ from i18n_check.utils import (
     config_i18n_src_file,
     config_invalid_keys_directories_to_skip,
     config_invalid_keys_files_to_skip,
-    config_invalid_keys_keys_to_ignore,
+    config_invalid_keys_regex_to_ignore,
     config_src_directory,
     filter_valid_key_parts,
     get_all_json_files,
@@ -97,7 +97,7 @@ def map_keys_to_files(
 
 def audit_i18n_keys(
     key_file_dict: Dict[str, List[str]],
-    keys_to_ignore_pattern: str = "",
+    keys_to_ignore_regex: str = "",
 ) -> Tuple[List[str], Dict[str, str]]:
     """
     Audit i18n keys for formatting and naming conventions.
@@ -107,8 +107,8 @@ def audit_i18n_keys(
     key_file_dict : Dict[str, List[str]]
         A dictionary where keys are i18n keys and values are lists of file paths where those keys are used.
 
-    keys_to_ignore_pattern : str, optional
-        A regex pattern to match keys that should be ignored during validation.
+    keys_to_ignore_regex : str, optional, default=""
+        A regex pattern to match with keys that should be ignored during validation.
         Keys matching this pattern will be skipped during the audit.
 
     Returns
@@ -118,15 +118,16 @@ def audit_i18n_keys(
         - A list of keys that are not formatted correctly.
         - A dictionary mapping keys that are not named correctly to their suggested corrections.
     """
-    # Filter out keys matching the ignore pattern
-    if keys_to_ignore_pattern:
-        filtered_key_file_dict = {
+    # Filter out keys matching the ignore pattern.
+    filtered_key_file_dict = (
+        {
             k: v
             for k, v in key_file_dict.items()
-            if not re.search(keys_to_ignore_pattern, k)
+            if not re.search(keys_to_ignore_regex, k)
         }
-    else:
-        filtered_key_file_dict = key_file_dict
+        if keys_to_ignore_regex
+        else key_file_dict
+    )
 
     invalid_keys_by_format = []
     invalid_keys_by_name = {}
@@ -303,7 +304,7 @@ key_file_dict = map_keys_to_files(
 )
 invalid_keys_by_format, invalid_keys_by_name = audit_i18n_keys(
     key_file_dict=key_file_dict,
-    keys_to_ignore_pattern=config_invalid_keys_keys_to_ignore,
+    keys_to_ignore_regex=config_invalid_keys_regex_to_ignore,
 )
 
 if __name__ == "__main__":